Update ISBN13 Conversion Adds small ISBN tests Fixes small issues Maintains search functionality
